Placed InfoListViewFragment inside the MainActivity
