+++ app/src/main/res/layout/fragment_info_list_view.xml
